+++ react/utils/constants.ts
@@ -6,6 +6,3 @@ export const B2B_BULK_IMPORT_BEST_PRACTICES_LINK = 'https://help.vtex.com/'
 // Permission codes
 export const ORGANIZATION_EDIT = 'buyer_organization_edit'
 export const ORGANIZATION_VIEW = 'buyer_organization_view'
-
-// Product codes
-export const B2B_LM_PRODUCT_CODE = '97'